requirements.txt
# ============================================================================
# CORE DEPENDENCIES
# ============================================================================
python-dotenv>=1.0.0
pydantic>=2.5.0
pydantic-settings>=2.1.0
email-validator>=2.1.0

# ============================================================================
# DATABASE
# ============================================================================
psycopg2-binary>=2.9.9
sqlalchemy>=2.0.23

# ============================================================================
# DATA PROCESSING
# ============================================================================
pandas>=2.1.4
numpy>=1.26.2

# ============================================================================
# LLM & AI
# ============================================================================
langchain-ollama>=0.1.0
langchain-core>=0.1.0

# ============================================================================
# WEB FRAMEWORK (Streamlit)
# ============================================================================
streamlit>=1.29.0

# ============================================================================
# UTILITIES
# ============================================================================
python-dateutil>=2.8.2
pytz>=2023.3

# ============================================================================
# LOGGING & MONITORING
# ============================================================================
colorlog>=6.8.0

# ============================================================================
# TESTING (Development)
# ============================================================================
pytest>=7.4.3
pytest-cov>=4.1.0
pytest-asyncio>=0.21.1
pytest-mock>=3.12.0
